@@ -69,7 +69,7 @@ func titleCaseWithAbbrHelper(abbrev []string) wordcase.Combiner {
 }
 
 // TitleCaseWithAbbr upper-cases the first letter of each word, or the whole word if it matches a given abbreviation
-func TitleCaseWithAbbr(abbrev interface{}, word string) (string, error) {
+func TitleCaseWithAbbr(abbrev any, word string) (string, error) {
 	a, l, err := helper.ListInfo(abbrev)
 	if err != nil {
 		return "", err
@@ -85,7 +85,7 @@ func TitleCaseWithAbbr(abbrev interface{}, word string) (string, error) {
 }
 
 // UppercaseFirst converts the first character in a string to uppercase
-func UppercaseFirst(s interface{}) string {
+func UppercaseFirst(s any) string {
 	str := fmt.Sprintf("%v", s)
 	if str == "" {
 		return ""
@@ -111,39 +111,39 @@ func Tab(n int) string {
 }
 
 // PadRight prints the given string in the given number of columns, right aligned
-func PadRight(n int, s interface{}) string {
+func PadRight(n int, s any) string {
 	f := fmt.Sprintf("%%-%dv", n)
 	return fmt.Sprintf(f, s)
 }
 
 // PadLeft prints the given string in the given number of columns, left aligned
-func PadLeft(n int, s interface{}) string {
+func PadLeft(n int, s any) string {
 	f := fmt.Sprintf("%%%dv", n)
 	return fmt.Sprintf(f, s)
 }
 
 // Bracket adds brackets around the given string
-func Bracket(item interface{}) string {
+func Bracket(item any) string {
 	return fmt.Sprintf("(%v)", item)
 }
 
 // QuoteSingle adds single quote around the given string
-func QuoteSingle(item interface{}) string {
+func QuoteSingle(item any) string {
 	return fmt.Sprintf("'%v'", item)
 }
 
 // QuoteDouble adds double quote around the given string
-func QuoteDouble(item interface{}) string {
+func QuoteDouble(item any) string {
 	return fmt.Sprintf("\"%v\"", item)
 }
 
 // QuoteBack adds back-quotes around the given string
-func QuoteBack(item interface{}) string {
+func QuoteBack(item any) string {
 	return fmt.Sprintf("`%v`", item)
 }
 
 // BracketWith adds brackets of a given type around the given string
-func BracketWith(bracketPair string, item interface{}) (string, error) {
+func BracketWith(bracketPair string, item any) (string, error) {
 	if len(bracketPair)%2 != 0 {
 		return "", fmt.Errorf("expected a set of brackets with matching left and right sizes")
 	}
@@ -239,7 +239,7 @@ func Suffix(suffix string, count int, content string) string {
 
 // ToColumn formats the given text to not take more than 'w' characters per line,
 // splitting on the space before the word that would take the line over.
-// If no space can be found, the line isn't split (ie words bigger than the line size are printed unsplit)
+// If no space can be found, the line isn't split (ie words bigger than the line size are printed un-split)
 func ToColumn(width int, content string) string {
 	var b strings.Builder
 	tail := ""
@@ -321,15 +321,15 @@ func SplitOn(glue string, content string) []string {
 }
 
 // TypeName returns the type of the given value as a string
-func TypeName(value interface{}) string {
+func TypeName(value any) string {
 	if value == nil {
 		return "nil"
 	}
 	return reflect.TypeOf(value).String()
 }
 
-// TypeKind returns the 'kind'' of the given value as a string
-func TypeKind(value interface{}) string {
+// TypeKind returns the "kind" of the given value as a string
+func TypeKind(value any) string {
 	if value == nil {
 		return "nil"
 	}
